+++ packages/core/src/codewhisperer/commands/startTransformByQ.ts
@@ -27,7 +27,6 @@ import {
     downloadHilResultArchive,
     findDownloadArtifactStep,
     getArtifactsFromProgressUpdate,
-    getTransformationPlan,
     getTransformationSteps,
     pollTransformationJob,
     resumeTransformationJob,
@@ -556,31 +555,6 @@ export async function pollTransformationStatusUntilPlanReady(jobId: string, prof
     }
     jobPlanProgress['generatePlan'] = StepProgress.Succeeded
     throwIfCancelled()
-}
-
-export async function openTransformationPlan(jobId: string, profile?: RegionProfile) {
-    let plan = undefined
-    try {
-        plan = await getTransformationPlan(jobId, profile)
-    } catch (error) {
-        // means API call failed
-        getLogger().error(`CodeTransformation: ${CodeWhispererConstants.failedToCompleteJobNotification}`, error)
-        transformByQState.setJobFailureErrorNotification(
-            `${CodeWhispererConstants.failedToGetPlanNotification} ${(error as Error).message}`
-        )
-        transformByQState.setJobFailureErrorChatMessage(
-            `${CodeWhispererConstants.failedToGetPlanChatMessage} ${(error as Error).message}`
-        )
-        throw new Error('Get plan failed')
-    }
-
-    if (plan !== undefined) {
-        const planFilePath = path.join(transformByQState.getProjectPath(), 'transformation-plan.md')
-        fs.writeFileSync(planFilePath, plan)
-        await vscode.commands.executeCommand('markdown.showPreview', vscode.Uri.file(planFilePath))
-        transformByQState.setPlanFilePath(planFilePath)
-        await setContext('gumby.isPlanAvailable', true)
-    }
 }
 
 export async function pollTransformationStatusUntilComplete(jobId: string, profile: RegionProfile | undefined) {
